+++ src/app/app.interface.ts
@@ -72,6 +72,31 @@ export type AppLinks = {
   medium?: string;
 };
 
+export type PresentationConfig = {
+  tabs: (
+    | { viewType: 'list'; label: string; groupIds: string[] }
+    | {
+        viewType: 'split';
+        label: string;
+        views: (
+          | {
+              viewType: 'list';
+              label: string;
+              groupIds: string[];
+            }
+          | {
+              viewType: 'split';
+              label: string;
+              views: {
+                label: string;
+                groupIds: string[];
+              }[];
+            }
+        )[];
+      }
+  )[];
+};
+
 export type AppDefinitionObject = {
   id: string;
   name: string;
@@ -79,6 +104,7 @@ export type AppDefinitionObject = {
   keywords?: string[];
   description: string;
   groups: Record<string, AppGroup>;
+  presentationConfig?: PresentationConfig;
   supportedNetworks: { [N in Network]?: AppAction[] };
   primaryColor?: string;
   url: string;