+++ src/apps/argo-finance/index.ts
@@ -1,0 +1,3 @@
+export { ARGO_FINANCE_DEFINITION, ArgoFinanceAppDefinition } from './argo-finance.definition';
+export { ArgoFinanceAppModule } from './argo-finance.module';
+export { ArgoFinanceContractFactory } from './contracts';