+++ src/apps/compound/helper/compound.borrow.contract-position-helper.ts
@@ -84,6 +84,7 @@ export class CompoundBorrowContractPositionHelper {
         // The amount borrowed can be derived simply by substracting the liquidity from the total supply
         // of tokens
         borrow: borrowLiquidity,
+        isActive: Boolean(borrowLiquidity > 0),
       };
       const borrowApy = appToken.dataProps.borrowApy;
 
@@ -94,8 +95,8 @@ export class CompoundBorrowContractPositionHelper {
       const images = appToken.displayProps.images;
       const statsItems = isNumber(borrowApy)
         ? [
-            { label: 'APY', value: buildPercentageDisplayItem(borrowApy) },
-            { label: 'Liquidity', value: buildDollarDisplayItem(dataProps.liquidity) },
+            { label: 'APY', value: buildPercentageDisplayItem(borrowApy * 100) },
+            { label: 'Liquidity', value: buildDollarDisplayItem(borrowLiquidity) },
           ]
         : [];