+++ src/apps/iron-bank/avalanche/iron-bank.supply.token-fetcher.ts
@@ -33,7 +33,6 @@ export class AvalancheIronBankSupplyTokenFetcher implements PositionFetcher<AppT
       getExchangeRate: ({ contract, multicall }) => multicall.wrap(contract).exchangeRateCurrent(),
       getSupplyRate: ({ contract, multicall }) => multicall.wrap(contract).supplyRatePerBlock(),
       getBorrowRate: ({ contract, multicall }) => multicall.wrap(contract).borrowRatePerBlock(),
-      getBorrowRateLabel: () => 'Borrow APY',
       getUnderlyingAddress: ({ contract, multicall }) => multicall.wrap(contract).underlying(),
       getExchangeRateMantissa: ({ underlyingTokenDecimals }) => underlyingTokenDecimals + 10,
     });