+++ src/apps/qi-dao/qi-dao.module.ts
@@ -5,6 +5,8 @@ import { QiDaoContractFactory } from './contracts';
 import { FantomQiDaoBalanceFetcher } from './fantom/qi-dao.balance-fetcher';
 import { FantomQiDaoFarmContractPositionFetcher } from './fantom/qi-dao.farm.contract-position-fetcher';
 import { FantomQiDaoVaultPositionFetcher } from './fantom/qi-dao.vault.position-fetcher';
+import { GnosisQiDaoBalanceFetcher } from './gnosis/qi-dao.balance-fetcher';
+import { GnosisQiDaoVaultPositionFetcher } from './gnosis/qi-dao.vault.position-fetcher';
 import { QiDaoVaultPositionBalanceHelper } from './helpers/qi-dao.vault.position-balance-helper';
 import { QiDaoVaultPositionHelper } from './helpers/qi-dao.vault.position-helper';
 import { PolygonQiDaoAnchorVaultPositionFetcher } from './polygon/qi-dao.anchor-vault.contract-position-fetcher';
@@ -33,6 +35,9 @@ import { QiDaoAppDefinition, QI_DAO_DEFINITION } from './qi-dao.definition';
     PolygonQiDaoVaultPositionFetcher,
     PolygonQiDaoYieldTokenFetcher,
     PolygonQiDaoAnchorVaultPositionFetcher,
+    // Gnosis
+    GnosisQiDaoBalanceFetcher,
+    GnosisQiDaoVaultPositionFetcher,
   ],
 })
 export class QiDaoAppModule extends AbstractApp() {}