+++ src/apps/synthetix/helpers/synthetix.synth.token-helper.ts
@@ -7,7 +7,7 @@ import { APP_TOOLKIT, IAppToolkit } from '~app-toolkit/app-toolkit.interface';
 import { buildDollarDisplayItem } from '~app-toolkit/helpers/presentation/display-item.present';
 import { getTokenImg } from '~app-toolkit/helpers/presentation/image.present';
 import { ContractType } from '~position/contract.interface';
-import { AppTokenPosition, ExchangeableAppTokenDataProps } from '~position/position.interface';
+import { AppTokenPosition } from '~position/position.interface';
 import { Network } from '~types/network.interface';
 
 import { SynthetixContractFactory } from '../contracts';
@@ -17,6 +17,11 @@ export type SynthetixSynthTokenHelperParams = {
   network: Network;
   resolverAddress: string;
   exchangeable?: boolean;
+};
+
+export type SynthetixAppTokenDataProps = {
+  exchangeable: boolean;
+  liquidity: number;
 };
 
 @Injectable()
@@ -55,17 +60,19 @@ export class SynthetixSynthTokenHelper {
         const price = Number(synthPrices[i]) / 10 ** 18;
         const pricePerShare = 1;
         const tokens = [];
+        const liquidity = supply * price;
 
         // Display Props
         const label = symbol;
         const secondaryLabel = buildDollarDisplayItem(price);
         const images = [getTokenImg(address, network)];
+        const statsItems = [{ label: 'Liquidity', value: buildDollarDisplayItem(liquidity) }];
 
-        const token: AppTokenPosition<ExchangeableAppTokenDataProps> = {
+        const token: AppTokenPosition<SynthetixAppTokenDataProps> = {
           type: ContractType.APP_TOKEN,
           appId: SYNTHETIX_DEFINITION.id,
           groupId: SYNTHETIX_DEFINITION.groups.synth.id,
-          network: network,
+          network,
           address,
           symbol,
           decimals,
@@ -74,12 +81,16 @@ export class SynthetixSynthTokenHelper {
           pricePerShare,
           tokens,
 
-          dataProps: { exchangeable },
+          dataProps: {
+            exchangeable,
+            liquidity,
+          },
 
           displayProps: {
             label,
             secondaryLabel,
             images,
+            statsItems,
           },
         };