+++ src/apps/vvs-finance/cronos/vvs-finance.farm-v2.contract-position-fetcher.ts
@@ -5,6 +5,7 @@ import { compact } from 'lodash';
 import { IAppToolkit, APP_TOOLKIT } from '~app-toolkit/app-toolkit.interface';
 import { Register } from '~app-toolkit/decorators';
 import { RewardRateUnit } from '~app-toolkit/helpers/master-chef/master-chef.contract-position-helper';
+import { getLabelFromToken } from '~app-toolkit/helpers/presentation/image.present';
 import { PositionFetcher } from '~position/position-fetcher.interface';
 import { ContractPosition } from '~position/position.interface';
 import { Network } from '~types/network.interface';
@@ -113,6 +114,8 @@ export class CronosVvsFinanceFarmV2ContractPositionFetcher implements PositionFe
           resolveTotalRewardRate: async ({ poolIndex }) =>
             poolRecords[poolIndex].rewards.map(r => r.rewardPerSecond).reduce((p, c) => p.add(c), BigNumber.from(0)),
         }),
+        resolveLabel: ({ stakedToken, rewardTokens }) =>
+          `Staked ${getLabelFromToken(stakedToken)} for ${rewardTokens.map(getLabelFromToken).join(', ')} farm`,
       });
 
     return positions;